+++ R/plot.R
@@ -360,8 +360,9 @@ PlotCatalog.SBS96Catalog <-
       count.cex <- cex
       for (i in 1:6) {
         j <- 16 + 16 * (i - 1)
-        k <- 1 + 16 * (i - 1)
-        text(bp[j], ymax * 1.15, labels = sum(catalog[k:(16 * i), ]),
+        k <- 1 + 16 * (i - 1)    
+        # Round the total mutation counts in case it is a reconstructed catalog
+        text(bp[j], ymax * 1.15, labels = round(sum(catalog[k:(16 * i), ])),
              adj = c(1, 1), xpd = NA, cex = count.cex)
       }
     }
@@ -562,7 +563,8 @@ PlotCatalog.SBS192Catalog <-
       for (i in 1 : 6) {
         j <- 32 + 32 * (i - 1)
         k <- 1 + 32 * (i - 1)
-        text(bp[j], ymax * 0.84, labels = sum(cat[k : (32 * i), 1]),
+        # Round the total mutation counts in case it is a reconstructed catalog
+        text(bp[j], ymax * 0.84, labels = round(sum(cat[k : (32 * i), 1])),
              adj = c(1, 1), xpd = NA, cex = cex)
       }
     }
@@ -1044,8 +1046,9 @@ PlotCatalog.DBS78Catalog <- function(catalog, plot.SBS12, cex,
     for (i in 1 : 10) {
       j <- c(9, 15, 24, 30, 39, 45, 51, 60, 69, 78)
       name <- substr(rownames(catalog), 1, 2)
+      # Round the total mutation counts in case it is a reconstructed catalog
       text(bp[j[i] + 0.5], ymax * 0.84, xpd = NA, cex = 0.8,
-           adj = c(1, 1), labels = sum(catalog[name == maj.class.names[i], ]))
+           adj = c(1, 1), labels = round(sum(catalog[name == maj.class.names[i], ])))
     }
   } else if (attributes(catalog)$catalog.type %in%
              c("counts.signature", "density.signature")) {
@@ -1646,7 +1649,8 @@ PlotCatalog.IndelCatalog <- function(catalog, plot.SBS12, cex,
       } else {
         counts[i] <- sum(catalog[(idx[i - 1] + 1):idx[i], 1])
       }
-      text(idx2[i], ymax * 0.8, labels = counts[i],
+      # Round the total mutation counts in case it is a reconstructed catalog
+      text(idx2[i], ymax * 0.8, labels = round(counts[i]),
            cex = 0.68, adj = 1, xpd = NA)
     }
   }